Extract class 'Storage'
