+++ src/main/java/Ui.java
@@ -1,0 +1,3 @@
+public class Ui {
+
+}
